Move ServiceBinder to top-level
