+++ simple-stack-example-mvp-fragments/src/main/java/com/zhuinden/simplestackdemoexamplefragments/util/scopedservices/ServiceProvider.java
@@ -3,6 +3,7 @@ package com.zhuinden.simplestackdemoexamplefragments.util.scopedservices;
 import android.support.annotation.NonNull;
 
 import com.zhuinden.simplestack.ScopedServices;
+import com.zhuinden.simplestack.ServiceBinder;
 
 public class ServiceProvider implements ScopedServices {
     @Override
