+++ simple-stack-example-mvp-view/src/main/java/com/zhuinden/simplestackdemoexamplemvp/util/scoping/ServiceProvider.kt
@@ -2,14 +2,15 @@ package com.zhuinden.simplestackdemoexamplemvp.util.scoping
 
 import com.zhuinden.simplestack.ScopeKey
 import com.zhuinden.simplestack.ScopedServices
+import com.zhuinden.simplestack.ServiceBinder
 import com.zhuinden.simplestackdemoexamplemvp.core.navigation.ViewKey
 
 class ServiceProvider: ScopedServices {
     interface HasServices: ScopeKey {
-        fun bindServices(serviceBinder: ScopedServices.ServiceBinder)
+        fun bindServices(serviceBinder: ServiceBinder)
     }
 
-    override fun bindServices(serviceBinder: ScopedServices.ServiceBinder) {
+    override fun bindServices(serviceBinder: ServiceBinder) {
         val key = serviceBinder.getKey<ViewKey>()
         if(key is HasServices) {
             key.bindServices(serviceBinder)
