+++ simple-stack-example-mvvm-fragments/src/main/java/com/zhuinden/simplestackexamplemvvm/util/ServiceProvider.java
@@ -4,6 +4,7 @@ import android.support.annotation.NonNull;
 
 import com.zhuinden.simplestack.ScopeKey;
 import com.zhuinden.simplestack.ScopedServices;
+import com.zhuinden.simplestack.ServiceBinder;
 
 public class ServiceProvider
         implements ScopedServices {
