+++ simple-stack-example-scoping/src/main/kotlin/com/zhuinden/simplestackexamplescoping/Utils.kt
@@ -7,7 +7,7 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import android.widget.Toast
-import com.zhuinden.simplestack.ScopedServices
+import com.zhuinden.simplestack.ServiceBinder
 import com.zhuinden.simplestack.navigator.Navigator
 
 /**
@@ -40,7 +40,7 @@ inline fun <reified T> Fragment.canFind(serviceTag: String = T::class.java.name)
 
 inline fun <reified T> Fragment.lookup(serviceTag: String = T::class.java.name) = Navigator.lookupService<T>(requireContext(), serviceTag)
 
-inline fun <reified T> ScopedServices.ServiceBinder.addService(service: T, serviceTag: String = T::class.java.name) {
+inline fun <reified T> ServiceBinder.addService(service: T, serviceTag: String = T::class.java.name) {
     add(serviceTag, service as Any)
 }
