+++ simple-stack/src/test/java/com/zhuinden/simplestack/ScopeLookupModeTest.java
@@ -28,12 +28,12 @@ import static org.assertj.core.api.Assertions.assertThat;
 public class ScopeLookupModeTest {
     private interface HasServices
             extends ScopeKey {
-        void bindServices(ScopedServices.ServiceBinder serviceBinder);
+        void bindServices(ServiceBinder serviceBinder);
     }
 
     private interface HasParentServices
             extends ScopeKey.Child {
-        void bindServices(ScopedServices.ServiceBinder serviceBinder);
+        void bindServices(ServiceBinder serviceBinder);
     }
 
     private static class ServiceProvider
@@ -70,7 +70,7 @@ public class ScopeLookupModeTest {
             }
 
             @Override
-            public void bindServices(ScopedServices.ServiceBinder serviceBinder) {
+            public void bindServices(ServiceBinder serviceBinder) {
                 if("parent1".equals(serviceBinder.getScopeTag())) {
                     serviceBinder.add("parentService1", parentService1);
                 } else if(name.equals(serviceBinder.getScopeTag())) {
@@ -103,7 +103,7 @@ public class ScopeLookupModeTest {
             }
 
             @Override
-            public void bindServices(ScopedServices.ServiceBinder serviceBinder) {
+            public void bindServices(ServiceBinder serviceBinder) {
                 if("parent2".equals(serviceBinder.getScopeTag())) {
                     serviceBinder.add("parentService2", parentService2);
                 } else if(name.equals(serviceBinder.getScopeTag())) {
@@ -407,7 +407,7 @@ public class ScopeLookupModeTest {
             }
 
             @Override
-            public void bindServices(ScopedServices.ServiceBinder serviceBinder) {
+            public void bindServices(ServiceBinder serviceBinder) {
                 if("parent1".equals(serviceBinder.getScopeTag())) {
                     serviceBinder.add("parentService1", parentService1);
                 } else if(name.equals(serviceBinder.getScopeTag())) {
@@ -440,7 +440,7 @@ public class ScopeLookupModeTest {
             }
 
             @Override
-            public void bindServices(ScopedServices.ServiceBinder serviceBinder) {
+            public void bindServices(ServiceBinder serviceBinder) {
                 if("parent2".equals(serviceBinder.getScopeTag())) {
                     serviceBinder.add("parentService2", parentService2);
                 } else if(name.equals(serviceBinder.getScopeTag())) {
@@ -512,7 +512,7 @@ public class ScopeLookupModeTest {
             }
 
             @Override
-            public void bindServices(ScopedServices.ServiceBinder serviceBinder) {
+            public void bindServices(ServiceBinder serviceBinder) {
                 if("parent1".equals(serviceBinder.getScopeTag())) {
                     serviceBinder.add("parentService1", parentService1);
                 } else if(name.equals(serviceBinder.getScopeTag())) {
@@ -545,7 +545,7 @@ public class ScopeLookupModeTest {
             }
 
             @Override
-            public void bindServices(ScopedServices.ServiceBinder serviceBinder) {
+            public void bindServices(ServiceBinder serviceBinder) {
                 if("parent2".equals(serviceBinder.getScopeTag())) {
                     serviceBinder.add("parentService2", parentService2);
                 } else if(name.equals(serviceBinder.getScopeTag())) {
@@ -650,7 +650,7 @@ public class ScopeLookupModeTest {
 
         TestKeyWithScope beep = new TestKeyWithScope("scope1") {
             @Override
-            public void bindServices(ScopedServices.ServiceBinder serviceBinder) {
+            public void bindServices(ServiceBinder serviceBinder) {
             }
         };
 
@@ -664,7 +664,7 @@ public class ScopeLookupModeTest {
             }
 
             @Override
-            public final void bindServices(ScopedServices.ServiceBinder serviceBinder) {
+            public final void bindServices(ServiceBinder serviceBinder) {
                 if(name.equals(serviceBinder.getScopeTag())) {
                     bindOwnServices(serviceBinder);
                 } else {
@@ -672,9 +672,9 @@ public class ScopeLookupModeTest {
                 }
             }
 
-            abstract void bindParentServices(ScopedServices.ServiceBinder serviceBinder);
+            abstract void bindParentServices(ServiceBinder serviceBinder);
 
-            abstract void bindOwnServices(ScopedServices.ServiceBinder serviceBinder);
+            abstract void bindOwnServices(ServiceBinder serviceBinder);
         }
 
         TestKeyWithExplicitParent boop = new TestKeyWithExplicitParent("scope2") {
@@ -685,11 +685,11 @@ public class ScopeLookupModeTest {
             }
 
             @Override
-            void bindParentServices(ScopedServices.ServiceBinder serviceBinder) {
+            void bindParentServices(ServiceBinder serviceBinder) {
             }
 
             @Override
-            void bindOwnServices(ScopedServices.ServiceBinder serviceBinder) {
+            void bindOwnServices(ServiceBinder serviceBinder) {
             }
         };
 
@@ -701,11 +701,11 @@ public class ScopeLookupModeTest {
             }
 
             @Override
-            void bindParentServices(ScopedServices.ServiceBinder serviceBinder) {
+            void bindParentServices(ServiceBinder serviceBinder) {
             }
 
             @Override
-            void bindOwnServices(ScopedServices.ServiceBinder serviceBinder) {
+            void bindOwnServices(ServiceBinder serviceBinder) {
             }
         };
